requirements.txt
httpx
click
jinja2
rich
python-whois